refactor(polytope_playground): extract console command registration into console module
Moves the ~690-line build_console (every wireframe/pole/section/space subcommand) out of main.rs into console.rs.
